fix alignment computation for large sets (#1558)
## Summary

Fix the alignment computation for large sets not matching the physical
alignment when using the C backend, which could cause heap/stack
corruption due to incorrectly computed aggregate type sizes.

## Details

Large sets (i.e., sets with more than 64 elements) are implemented as
arrays of `uint8` by the C code generator, which have an alignment of 1
byte, but `sizealignoffsetimpl` considered the sets to be aligned
the same as `int`.

Since the computed alignment is used when computing internal field
offsets and aggregate type sizes, it was possible for `offsetof` and
`sizeof` to report *larger* sizes and offsets than what the actual
values are, thus being able to cause out-of-bound reads/writes when
used in conjunction with low-level memory operations (e.g., `copyMem`).