sempass2: fix bug with `.closure` inference (#1560)
## Summary

Fix inference failure and the compiler subsequently crashing for
forwarded inner procedures and generic inner routines.

## Details

`.closure` inference for didn't consider forwarded routines properly
and generic owners were also not ignored, leading to some the affected
inner routines not being inferred as requiring the `.closure` calling
convention, which then led to an internal assertion failure during
lambda-lifting.

Still-forwarded routines don't have a body to analyze at the time of
usage, and thus have to be assumed to capture, unless an explicit
calling convention is specified that's not `.closure`.