transf: don't patch `result` usage in suspend blocks
For the `result` variable to work as expected within suspend blocks, its
usage cannot simply be replaced with the duplicate symbol, as that would
break result assignments or modifications.

Instead, the `result` symbol is now kept as is, with `mirgen` emitting
assigning `result` to the copy created of the result variable's value
prior to the fork.