continuations: better detection of `result` usage
The `result` variable is also used implicitly when a routine returns,
which wasn't considered till now, leading to the `result` variable not
being saved in cases where it should have been.

Detecting graph exits is a rather blunt solution, as it also triggers
for exceptional exits (where the result variable is not used), meaning
that the result variables is sometimes saved even when it doesn't
have to be.