net: do not reuse socket after failed connect
The socket state after a failed connect attempt is unspecified in POSIX.
While on certain OS like Linux or Windows, the socket can be reused and
another `connect` attempt could be made, others will no longer accept
the socket for future operations. This caused macOS CI to fail
sporadically when a test domain (example.com), failed to connect on
first try.

This commit removes the socket reuse mechanism entirely, given that this
is recommended by POSIX and Linux manuals for portable code, as well as
a lack of evidence that this optimization provides adequate performance
benefits to justify maintaining it for supported platforms.

Small improvements on resource cleanup and error messages are also
applied as part of this change.