sempass2: fix crash with generic .closure routine (#1570)
## Summary

Fix the compiler crashing when using the instance of an inner generic
.closure routine in a routine that's not a parent of the generic
routine.

## Details

Detection in `sempass2` of whether a capture is used the `owner` of the
first-appearing outer symbol as the traversal target. If the outer
symbol is that of a generic instantiation, the owner is the
originating-from generic routine, and since the traversal in
`canCaptureFrom` skips generic owners (as it should), this meant that
traversal skipped over the target and ended up fetching the owner of
the top-level `skPackage` symbol (which is `nil`), causing a nil
pointer access.

The generic owner is now skipped properly.