transf: fix indirect tail calls not being detected as such
Fix calls where the callee is a `.tailcall` procval not being
considered a tail call when the callee type is a generic type
instantiation.

Type skipping in `transf` was missing for the relevant "is callee a
`.tailcall` proc type" checks, leaving such as calls as normal calls.

A MIR-output-based test is added to ensure the calls in question being
tail calls.