transf: fix indirect tail calls not being detected as such (#1572)
## Summary

Fix calls where the callee is a `.tailcall` procval not being
considered a tail call when the callee type is a generic type
instantiation.

## Details

Type skipping in `transf` was missing for the relevant "is callee a
`.tailcall` proc type" checks, leaving such calls as normal calls.

A MIR-output-based test is added to ensure the calls in question being
tail calls.