fix miscompilation with indirect tail calls (#1573)
## Summary

Fix a bug with the tail call elimination that caused target compiler
errors, or run-time errors, with the JS and C targets, in some edge
cases involving indirect tail calls.

## Details

In the MIR type IR, there are no `sink` types, meaning that
`proc(x: sink T)` and `proc(x: T)` map to the same type
representation -- the one first registered with the environment is
considered the canonical one.

The tail call elimination looked up the `PType` for the canonical type
of `.tailcall` procvals, yielding a `PType that's not the original for
the MIR type that's not the canonical one. Since the `sink` modifier is
significant for the parameter tuple type creation, this led to the
tuple having the wrong type for one of the indirect calls.

The lookup is changed to use the original instead of the canonical MIR
type, fixing the problem. A regression test reproducing the problem
is added.