fix heap-cell cleanup calling the wrong destructor (#1587)
## Summary

Fix `ref`s to objects calling the wrong destructor upon the refcount
reaching zero in some narrow circumstances involving generic types and
`sink` type modifiers.

## Details

Due to the sighash-based canonicalization used prior to type-bound op
lifting ignoring `sink` modifiers, the same hook was lifted for both
`ref Obj[proc(x: sink T)]` and `ref Obj[proc(x: T)]`, which led to the
wrong destructor being called on cleanup when the object type is final.

In a very narrow situation, where all of the following apply:
* the `ref` type itself is generic
* the object type is never used outside the `ref` type
* the object type has no user-defined destructor, only an implicit one
* the object type is non-final
* the types reach the MIR phase in the wrong order
the canonicalization led to the object not having its destructor called
at all, due to the destructor not being lifted, thus *leaking memory*
if the object type requires destruction for some of its fields.

Two types for which `sameType(a, b)` results in `false` should produce
different sighashes, and so hasing is changed to not ignore `sink`
types, fixing both issues.

RTTI creation in `ccgtypes` relied on `sink` not contributing to
sighashes; both `genTypeInfoV1` and `genTypeInfoV2` are changed to
manually skip `sink`, so that `sink T` and `T` continue to use the same
RTTI.

In addition, the scanning for types to lift the ops for in `sempass2`
is fixed to handle generic ref types used as the constructor in object
construction expressions, which makes the code work as intended.