mirtypes: remove type indirections (#1596)
## Summary

Simplify the MIR type IR by removing the `tkIndirect` pseudo type kind.
This is an internal-only change.

## Details

The `tkIndirect` MIR type kind was originally to allow for fully
handling type skipping, but this turned out to not be useful in
practice, actually making MIR type traversal worse, by introducing
the need to "skip" types to the IR.

The pseudo type kind is removed, with type translation changed such
that wrapper types irrelevant at the MIR level are translated to
type symbols mirroring the symbol of their skipped type.

In practice, this means that querying the canonical or lowered
representation for a type symbol always yields the "real" description
now, not an indirection, eliminating the need to skip the indirection.
None of the usage sites are adjusted, however.