remove the sink inference feature (#1595)
## Summary

Remove the compiler's unsound sink inference feature. This includes the
`sinkinference` switch and pragma, as well as the `.nosinks` pragma.

## Details

Modifying procedure signatures after they already participated in type
inference and checking (which is what sink inference did) is unsound,
as it invalidates the results of previous analysis.

Even if sink inference were implemented in a way such that it doesn't
make unsound modifications to routine signatures, it would still have
the problems that:
1. it silently changes the operational semantics at a routine's
   callsite
2. it changes concrete (i.e., non generic) routine signatures to
   something incompatible (type wise), allowing for confusing type
   mismatch errors

Everything related to the feature is removed from the compiler:
* the sink inference implementation
* the `sinkinference` compiler switch and option
* the `.nosinks` pragma
* the feature's documentation

The compiler also no longer advertises the sink inference capability
(`nimHasSinkInference`). All mentions and usages of sink inference
are removed from both the compiler and system/standard library.

### Impact

Sink inference was previously disabled by default, with it being
selectively enabled for `system` and a few standard library modules
(i.e., the ones including `inclrtl`).

The feature being gone means that `sink` annotations aren't inferred
anymore for the aforementioned modules.

### Additional Adjustments

* remove `.nosinks` annotations from tests
* adjust tests relying on sink inference
* explicitly mark unjoinable tests as such. They were previously
  prevented from being joined into by the local `nim.cfg` file
* add explicit `sink` annotations to some parameters in `system` where
  needed for proper cleanup