std: remove deprecation for `macros.callsite`
`macros.callsite` was marked deprecated and is no longer deprecated, because it's still very useful.

## Details

It was initially thought that the main use case for `macros.callsite` was fetching arbitrary arguments to a macro call. That use case was taken over by `varargs[untyped]`, but this is obvious short sighted. The routine still has utility and is no longer marked deprecated.