strs_v2: improve performance of string operations (#1602)
## Summary

Make growing a string always initialize new characters to NUL (`'\0'`) 
and improve performance of string operations that resize the internal
buffer.

## Details

So far, string payloads were always allocated with `alloc0`, filling
the whole allocation with zero. In most cases, this is unnecessary,
as either parts of or the whole buffer is overwritten immediately
afterwards.

The payload is now generally allocated without zeroing it, with
memory regions only being filled with zero when they become active
without explicit initialization (i.e., `newString`, or when growing
a string via `setLen`).

This also makes the behaviour of growing strings via `setLen`
consistent across backends and with `seq` (where new elements are
default initialized). A test is added to codify this.