sigmatch: remove lambda analysis workaround
Since a lambda expression analysis error led to the candidate being
rejected, the error had to be reported eagerly, otherwise the error
message would have been confusing, with no indication of the error part
of the lambda expression being the underlying problem.

Since a faux match is now possible in this case, eagerly reporting is
no longer necessary.