pragmas: replace `nkError`-based implicit pragma culling (#1607)
## Summary

Change culling of non-applicable implicit pragmas to work without
creating and discarding `nkError` nodes.

## Details

Creating an `nkError` node and then not committing to it is unusual and
will also cease to work once creating an `nkError` node means "emit
diagnostic". Handling of implicit pragmas made use of `nkError` to
detect and omit pragmas that are not applicable in the current context.

The relevant procedures are changed such that handling an invalid
(read, non- applicable) pragma is done via a callback (effect handler).

Empty nodes returned from the callback are silently discarded, which
implicit pragma processing uses to dismiss non-applicable pragmas.
Pragma processing in other contexts stays unchanged, as it continues
to turn non-applicable pragmas into errors.