diff: greatly speed up `myersDiff` (#1611)
## Summary

Significantly improve performance of `diff.myersDiff`, especially for
sequences with large differences.

## Details

For input sequences with a large number of differences, the procedure
was very slow, due to the explicit `seq` copy it created for every
search point.

The implementation is changed to use a single sequence for representing
the graph, with the search points then pointing into the sequence.
This:
* gets rid of the `seq` copies
* reduces the in-memory size of the table entries
* reduces the number of allocations