mirtypes: rename `tkRecord` to `tkStruct`
Summary
=======

Changed the name of the `tkRecord` MIR type kind to the better fitting
`tkStruct` and update the name of related routines.

Details
=======

In the context of type theory, 'record' usually refers to an unordered
set of fields, which what was previously called record types in the
MIR is not -- the fields in a MIR record type are ordered.

'struct' (from structure) better describes what the type represents.