sem: codegendecl pragma works for generic routines (#1623)
## Summary

`codegendecl` now works on generic routines.

## Details

`codegendecl`  pragma processing no longer corrupts the pragma AST. The
processing error meant asyncjs' use of  `codegendecl`  with generic
routines mutated the pragma AST position of the annotated routine
definition on the first pass, upon generic instantiation the pragma was
processed again but this time it was malformed, resulting in misleading
errors.

Besides fixing this behaviour, by not unnecessarily mutating the AST in
the first place, a test was added to prevent future regressions.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>