tssl: ignore SslError in "close during write" test
The goal of the test is to verify that closing a SSL socket
after a fatal error happened will not cause the action to raise
an error. Sometimes this fatal error manifests as `SslError`,
which was not handled, causing intermittent failures (on Windows).

This commit modifies the test to also ignore `SslError` should that
happens, since we are only interested in the `close` statement not
raising an error.