tssl: ignore SslError in "close during write" test (#1633)
## Summary
Fix intermittent Windows CI failures from an unhandled  `SslError` 
being raised due to abrupt peer disconnection in  `tssl` .

## Details
The goal of the test is to verify that closing an SSL socket after a
fatal error happens will not cause the action to raise an error.
Sometimes this fatal error manifests as  `SslError` , which was not
handled, causing intermittent failures (on Windows).

This commit modifies the test to also ignore  `SslError`  should that
happen, since we are only interested in the  `close`  statement not
raising an error.