liftdestructors: generate proper `isNil` check (#1630)
## Summary

Fix an internal issue with type-bound operator lifting for closure
types.

## Details

Instead of a proper  `isNil`  check, generation of the "not nil" check
for
closure environments used the pointer expression itself as the
condition expression of an `if`.

Since the expression was translated as is and a pointer value is valid
in an `if` condition position in C, this worked, but it's not correct
according to NimSkull's type system. A proper "not nil" check is
generated now.

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>