utils: make MIR pretty-printing output more stable (#1632)
## Summary

Pretty-print the anonymous constants as their value in `--expandArc`
and `--showir` output.

## Details

When the MIR environment is available, anonymous constants are now
rendered as their value, not as their raw ID (which is dependent on
externalities unrelated to the body of the procedure).

The goal is to keep tests inspecting `--expandArc` and `--showir`
output from starting to fail when unrelated processing changes in
the future -- the tests affected by the change in rendering are
updated.