mirtypes: properly handle recursive imported types
In the MIR type IR, imported types use a dedicated type kind
(`tkImported`), which has the "to be treated as" type as the element/
base.

Due to unconditionally updating the `PType`->MIR mapping when
translating `importc`ed object types meant that usages of the imported
types within its own body (via a pointer/ref indirection) used the
`tkStruct` symbol, not the `tkImported` symbol.

```nim
type Foreign {.importc: "...".} = object
  field: ptr Foreign
  # ^^ the field had type `(tkPtr (tkStruct ...))`,
  # not `(tkPtr (tkImported ...))`
```

To fix this, `handleImported` now adds a symbol and registers a
mapping right away, with creation of a MIR type symbol for the
"internal" type of `importc`ed types leaving the mapping as is.