mir: distinguish between procedure and pointer-to-procedure (#1641)
## Summary

Add the concept of procedure signatures to the MIR type IR and
translate `tyProc` types used in valued contexts to pointer-to-
procedure types.

## Details

* add `mirtypes.addSignature`, for registering `tyProc` types to be
  representing procedure signatures
* translate non-`.closure` `tyProc` types to pointer-to-procedure types
  in `mirtypes.typeToMir`
* adjust `.tailcall` procedure detection in `tailcall_elim` to handle
  pointer-to-procedure types
* change `ccgtypes` to translate `tkProc` types to C proc types
  * the `N_<name>_PTR` C macros in `nimbase.h` are now obsolete, but
    have to be kept for bootstrapping reasons
* adjust the `.closure` lowering in `cgen` to always return a
  pointer-to-procedure type

Not all target languages distinguish between procedure values and
pointer-to-procedure values (only C does), but for those that do, code
generation becomes simpler when there's a distinction made at the MIR
type level.