fix terminal expression used as construction operand (#1643)
## Summary

Fix using a terminal expressions as an object construction operand
resulting in incorrect run-time behaviour.

## Details

Elimination of unreachable expression didn't handle object constructors
properly, not adding the terminal expression to the resulting statement
list. For example:

```nim
  Obj(a: x, b: (stmt; return; 1))
  # was turned into just
  discard x
  # `stmt; return` was erroneously omitted 
```

Since the surrounding statement was marked as not returning, the lack
of terminator statement caused incorrect MIR to be produced by
`mirgen`.

The terminal expression is now added to the produced statement list,
and a test is added to make sure eliminating construction expressions
works.