mirtypes: properly translate recursive imported types (#1642)
## Summary

Fix an internal issue with MIR type translation of recursive
imported types.

## Details

In the MIR type IR, imported types use a dedicated type kind
(`tkImported`), which has the "to be treated as" type as the element/
base.

Due to unconditionally updating the `PType`->MIR mapping when
translating `importc`ed object types, usages of the imported types
within their own body (via a pointer/ref indirection) used the
`tkStruct` symbol, not the `tkImported` symbol.

```nim
type Foreign {.importc: "...".} = object
  field: ptr Foreign
  # ^^ the field had type `(tkPtr (tkStruct ...))`,
  # not `(tkPtr (tkImported ...))`
```

This led to internal type mismatches between the actual and computed
type of MIR access expression, but since the type of access expressions
is never recomputed at the moment, the bug had no user-facing effect. 

To fix the bug, `handleImported` now adds a symbol and registers a
mapping right away, with creation of a MIR type symbol for the
"internal" type of `importc`ed types leaving the mapping as is.