mirpasses: fix the "move unscoped" pass
An `if` followed by a `scope` can still "contain" defs who are not
"scoped", as it's possible for there to still be defs between the end
of the scope and the end of the `if`.

Aside: using a pseudo basic-block representation instead of a real one
was a major mistake. Oh well.