sem: fix illegal proc type conversions being allowed (#1670)
## Summary

Fix conversion between unrelated `proc` type being allowed when their
only difference was in their tag or exception effects.

## Details

Judging whether a conversion is allowed (`checkConvertible`) uses
`sameType` first, which so far ignored effects on proc types. This led
to proc types that are equal in everything except their effects to be
admitted for conversion, even when the effects were incompatible.

To fix the problem, `sameType` is changed to also consider effect list
for proc types, but only when the `IgnoreTupleFields` flag is not
provided (which is already used to disable `.noSideEffect` being
considered).