mirflow: implement folding without recursion
The tree structure the statement list may form can get very deeply
nested, especially when high-level case statements with many branches
are involved, which can lead to stack overflow or exceeding the maximum
recursion depth.

Folding is now done without recursion and as part of constructing the
block structures, which also gets rid of the having to do a separate
pass over the list.